Fix non-graceful server shutdowns and improve signal handling (#41)
* fix: replace os.system() with subprocess.call()

https://docs.python.org/3/library/subprocess.html#replacing-os-system
os.system() ignores SIGINT, which is the stop signal set by the Dockerfile.
This fix should immediately kill the Reforger process upon being asked to stop.

* feat: send SIGTERM to reforger when stopping

This provides a chance for the Reforger process to cleanup.

* feat: kill Reforger process on unhandled exception

If an unknown exception or a second interrupt occurs,
ensure the process is killed.

* feat: propagate exit code from Reforger process

* feat: handle SIGTERM on launch.py

This allows omitting STOPSIGNAL from Dockerfile.

* fix: use shlex.join() when printing command

This ensures the command can actually be invoked by the user
in case some of the arguments contain whitespace or other
unsafe characters.

* fix: stop reforger using SIGINT instead of SIGTERM

* refactor: explicitly skip kill clause when raising SystemExit

The purpose of this is making the program flow more obvious,
since proc.kill() already does nothing on completed processes.